requirements.txt
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.10.1
autopep8==2.0.2
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.5
cryptography==40.0.1
DateTime==5.1
distlib==0.3.6
dnspython==2.3.0
fastapi==0.95.0
filelock==3.8.2
Flask==2.1.3
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
frozendict==2.3.7
greenlet==1.1.2
h11==0.14.0
html5lib==1.1
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
multitasking==0.0.11
numpy==1.24.2
packaging==23.0
pandas==2.0.0
pipenv==2022.12.19
platformdirs==2.6.0
plotly==5.14.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.39
starlette==0.26.1
tenacity==8.2.2
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.17.1
virtualenv-clone==0.5.7
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.1
Werkzeug==2.1.2
yfinance==0.2.17
zope.interface==6.0
